pyproject.toml
[project]
name = "reverb-alerts"
version = "0.1.0"
description = "CLI tool to monitor Reverb.com for gear deals below configured price thresholds"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "firecrawl-py",
    "pyyaml",
    "click",
    "pydantic-ai",
    "pydantic",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/reverb_alerts"]

[project.scripts]
reverb-alerts = "reverb_alerts.cli:main"
